Ensures the implementations do raise an error when they should.
The following functions must return valid results (if it makes sense for the game) for valid non-chance players (i.e. 0 <= player < num_players_).

- std::string InformationStateString(Player player)
- void InformationStateTensor(Player player, std::vector<double>* values)
- std::vector<double> InformationStateTensor(Player player)
- std::string ObservationString(Player player)
- void ObservationTensor(Player player, std::vector<double>* values)
- std::vector<double> ObservationTensor(Player player)

These functions should raise an error on any invalid player. In particular, the equivalent functions without the `Player` argument should return an error on chance and terminal nodes.

The files that contains real changes are:
- spiel.h (with added documentation)
- api_tests.py (added test)
- basic_tests.py (some documentation)

The rest are the fixes to conform to the API expectations.

PiperOrigin-RevId: 306225343
Change-Id: Iec449c47c14fea9a6409bc01528633e3ca638be8