Implementation of the deep online mirror descent algorithm.
munchausen_deep_mirror_descent.py contains a MunchausenDQN agent, which is a
variant of the DQN agent with Munchausen trick to avoid storing the previous
policies and distributions, and a DeepOnlineMirrorDescent class that
encapsulates the training of MunchausenDQN agents and updating the policy and
distribution after each OMD iteration. This simplifies the logic in the runner
(e.g. in mfg_munchausen_domd_jax.py example).

PiperOrigin-RevId: 428025986
Change-Id: I79063bfa124ab342fecca6e3068945a26d2bb1b5