Adding Spades
Spades implementation is largely based off of the Bridge implementation. Runs a single round and returns the points earned/lost from the round (not the overall team's score). Note that currently the parameters use_mercy_rule, mercy_threshold, and win_threshold are not actually used at the moment. They will most likely be moved to and managed by the training script. However, score_partnership_0 and score_partnership_1 are needed in order to determine if a bag penalty is earned from the round.